last changes baby ohhhhhhhhhhh 🚀🚀🚀🚀🚀

+++ includes/push_swap.h
@@ -6,7 +6,7 @@
 /*   By: flverge <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/25 11:55:18 by flverge           #+#    #+#             */
-/*   Updated: 2024/01/24 18:39:23 by flverge          ###   ########.fr       */
+/*   Updated: 2024/01/25 09:44:11 by flverge          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,10 +31,10 @@ typedef struct s_node
 
 typedef struct s_algo
 {
-	int	max;
-	int	soustraction;
-	int	num_lower;
-	int	num_upper;
+	size_t	max;
+	size_t	soustraction;
+	size_t	num_lower;
+	size_t	num_upper;
 }			t_algo;
 
 // init_stack
